+++ changelog.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## Upcoming
+## 3.18.0
 * Critical is now an Attribute modifier
   * Players no longer have 5% chance to crit
-  * `iguanatweaksreborn:critical_chance` defines the chance for a crit to happen (between 0 and 1), defaults to no chance
+  * `iguanatweaksreborn:critical_chance` defines the chance for a crit to happen (between 0 and 1), defaults to 0
   * `iguanatweaksreborn:critical_damage` defines the damage multiplier for a crit. Defaults to 0.5 (+50%)
 * Anvil Repair
   * Repairing Items in an anvil now costs ~20% less